GitStory-AI/requirements.txt
gitpython
matplotlib
moviepy
pillow